+++ NuniToolbox/Enum/EnumTranslationType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace NuniToolbox.Enum
+{
+    public enum EnumTranslationType : byte
+    {
+        Default,
+        Singular,
+        Plural,
+        Abbreviation
+    }
+}